+++ CHANGELOG.md
@@ -23,6 +23,9 @@
   - `column.headerCellClass`
   - `column.formatterOptions`
     - More info in [#2104](https://github.com/adazzle/react-data-grid/pull/2104)
+  - `column.editor2`
+  - `column.editorOptions`
+    - More info in [#2102](https://github.com/adazzle/react-data-grid/pull/2102)
   - `scrollToRow` method
     - ⚠️ This replaces the `scrollToRowIndex` prop
 - **Removed:**